requirements.txt
pytest
pyyaml
pathspec
build
twine
pyinstaller
pytest-cov